refactor: QueryDsl -> JdbcTemplate 으로 변경  - 단순 Count 쿼리 JdbcTemplate로 리팩토링

+++ dailyge-api/src/main/java/project/dailyge/app/core/retrospect/presentation/RetrospectReadApi.java
@@ -28,7 +28,7 @@ public class RetrospectReadApi {
         @OffsetPageable final CustomPageable page
     ) {
         final AsyncPagingResponse<RetrospectJpaEntity> retrospectByPage = retrospectReadService.findRetrospectAndTotalCountByPage(dailygeUser, page);
-        final long totalPageCount = page.getTotalPageCount(retrospectByPage.totalCount());
+        final int totalPageCount = page.getTotalPageCount(retrospectByPage.totalCount());
         final RetrospectPageResponse payload = new RetrospectPageResponse(retrospectByPage.data(), totalPageCount);
 
         return ApiResponse.from(OK, payload);
